chore: Rename Project to transformer tf finetune
