+++ .circleci/config.yml
@@ -2,7 +2,7 @@ version: 2.1
 
 executors:
   python-executor:
-    working_directory: ~/transformers-bart-finetune
+    working_directory: ~/transformers-tf-finetune
     docker:
       - image: circleci/python:3.8
 
@@ -36,7 +36,7 @@ jobs:
           name: Run Tests
           command: |
             source env/bin/activate
-            pytest --cov-branch --cov transformers_bart_finetune
+            pytest --cov-branch --cov transformers_tf_finetune
 
       - run:
           name: Upload coverage report
@@ -53,8 +53,8 @@ jobs:
           name: Run black, isort
           command: |
             source env/bin/activate
-            black --check transformers_bart_finetune scripts tests
-            isort transformers_bart_finetune scripts tests
+            black --check transformers_tf_finetune scripts tests
+            isort transformers_tf_finetune scripts tests
 
 workflows:
   main:
