+++ README.md
@@ -1,8 +1,8 @@
-# Transformers BART finetune
+# Transformers tf finetune
 
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 [![Imports: isort](https://img.shields.io/badge/%20imports-isort-%231674b1?style=flat&labelColor=ef8336)](https://pycqa.github.io/isort/)
-[![cosmoquester](https://circleci.com/gh/cosmoquester/transformers-bart-finetune.svg?style=svg)](https://app.circleci.com/pipelines/github/cosmoquester/transformers-bart-finetune)
-[![codecov](https://codecov.io/gh/cosmoquester/transformers-bart-finetune/branch/master/graph/badge.svg?token=GTsIlZy6oG)](https://codecov.io/gh/cosmoquester/transformers-bart-finetune)
+[![cosmoquester](https://circleci.com/gh/cosmoquester/transformers-tf-finetune.svg?style=svg)](https://app.circleci.com/pipelines/github/cosmoquester/transformers-tf-finetune)
+[![codecov](https://codecov.io/gh/cosmoquester/transformers-tf-finetune/branch/master/graph/badge.svg?token=GTsIlZy6oG)](https://codecov.io/gh/cosmoquester/transformers-tf-finetune)
 
-Script to train hugginface transformers BART with Tensorflow 2
+Script to train hugginface transformers models with Tensorflow 2
