+++ setup.py
@@ -4,12 +4,12 @@ with open("README.md") as f:
     long_description = f.read()
 
 setup(
-    name="transformers-bart-finetune",
+    name="transformers-tf-finetune",
     version="0.0.1",
-    description="Script to train hugginface transformers BART with Tensorflow 2",
+    description="Script to train hugginface transformers models with Tensorflow 2",
     python_requires=">=3.6",
     install_requires=["tensorflow>=2", "tensorflow-addons", "transformers"],
-    url="https://github.com/cosmoquester/transformers-bart-finetune.git",
+    url="https://github.com/cosmoquester/transformers-tf-finetune.git",
     author="Park Sangjun",
     classifiers=[
         "Programming Language :: Python :: 3",
